+++ Cargo.toml
@@ -15,3 +15,4 @@ rust-code-analysis = "0.0.24"
 rayon = "1.5.1"
 git2 = "0.13.23"
 globset = "0.4.8"
+term-table = "1.3.2"
